@@ -1,39 +1,44 @@
 {
   "name": "fruits_in_choco_frontend",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@apollo/client": "^3.6.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "bootstrap": "^5.0.2",
-    "node-sass": "^6.0.1",
+    "@apollo/client": "^3.7.1",
+    "@emotion/react": "^11.10.5",
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@mui/lab": "^5.0.0-alpha.107",
+    "@mui/material": "^5.10.13",
+    "@mui/styles": "^5.10.10",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.1.3",
+    "bootstrap": "^5.2.2",
+    "material-table": "^1.68.1",
+    "net": "^1.0.2",
+    "node-sass": "^7.0.3",
     "notistack": "^1.0.10",
-    "prop-types": "^15.7.2",
-    "qs": "^6.10.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0-beta.4",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.0",
+    "prop-types": "^15.8.1",
+    "qs": "^6.11.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.6.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.4.3",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
-    "redux-form": "^8.3.7",
-    "redux-thunk": "^2.3.0",
-    "reselect": "^4.0.0",
-    "sematable": "^2.9.0",
-    "sockjs-client": "^1.6.0",
+    "redux-form": "^8.3.8",
+    "redux-thunk": "^2.4.2",
+    "reselect": "^4.1.7",
+    "sockjs-client": "^1.6.1",
     "stompjs": "^2.3.3",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^3.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
