@@ -3,7 +3,7 @@ import style from './EditCategoryPage.module.scss';
 import {reduxForm} from "redux-form";
 import {connect, useDispatch} from "react-redux";
 import {CategoryForm} from "../common/CategoryForm";
-import {useHistory} from "react-router-dom";
+import {useLocation, useNavigate} from "react-router-dom";
 import {loadCategoryById, updateCategoryThunk} from "../../../../redux/thunks/category_thunks";
 
 let EditCategoryReduxForm = reduxForm({form: 'edit_category', enableReinitialize: true})(CategoryForm);
@@ -17,10 +17,10 @@ EditCategoryReduxForm = connect(
 
 const EditCategoryPage = (props) => {
     const dispatch = useDispatch();
-    const history = useHistory();
+    const {pathname} = useLocation();
 
     useEffect(() => {
-        dispatch(loadCategoryById(history.location.pathname.split('/').pop()));
+        dispatch(loadCategoryById(pathname.split('/').pop()));
     }, []);
 
     const onSubmit = (formData) => {
