@@ -2,16 +2,16 @@ import React, {useEffect} from "react";
 import Preloader from "../../common/Preloader/Preloader";
 import {ORDER_STATUS_CONFIRMED, ORDER_STATUS_DECLINED, ORDER_STATUS_NOT_CONFIRMED} from "../../utils/constants";
 import {useDispatch, useSelector} from "react-redux";
-import {useHistory} from "react-router-dom";
+import {useLocation, useNavigate} from "react-router-dom";
 import {loadOrderById, updateOrderThunk} from "../../../redux/thunks/order_thunks";
 
 const OrderInfo = (props) => {
     const dispatch = useDispatch();
-    const history = useHistory();
+    const {pathname} = useLocation();
     const order = useSelector(state => state.orderReducer.order);
 
     useEffect(() => {
-        dispatch(loadOrderById(history.location.pathname.split('/').pop()));
+        dispatch(loadOrderById(pathname.split('/').pop()));
     }, [])
 
 
