@@ -5,16 +5,16 @@ import {faCartPlus, faCheckCircle} from "@fortawesome/free-solid-svg-icons";
 import {isProductInCart} from "../../utils/localStorageFunctions";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {useDispatch} from "react-redux";
-import {useHistory} from "react-router-dom";
+import {useNavigate} from "react-router-dom";
 import {saveProductToCart} from "../../../redux/thunks/cart_thunks";
 
 const ProductCard = ({product}) => {
     const dispatch = useDispatch();
-    const history = useHistory();
+    const navigate = useNavigate();
 
     const selectProduct = (e) => {
         e.preventDefault();
-        history.push({pathname: `/products/${product.id}`})
+        navigate(`/products/${product.id}`)
     };
 
     const cartButtonClickHandler = (e) => {
